+++ core/src/action.rs
@@ -3,28 +3,25 @@ pub mod error {
     pub(crate) use anyhow::{anyhow, bail, ensure, Context as _};
     pub use anyhow::{Error, Result};
 }
-use std::collections::HashMap;
-use std::path::Path;
-use std::sync::Arc;
-use std::time::Duration;
+
+use std::{path::Path, time::Duration};
 
 use chrono::{DateTime, Local};
-use colored::{Color, Colorize};
-use crossterm::terminal;
-use error::*;
+use colored::Colorize;
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
-use kpr_webclient::problem_id::ProblemGlobalId;
-use kpr_webclient::{PgLang, ProblemInfo, SampleTestcase, Url};
-use tokio::sync::Mutex;
+use kpr_webclient::{problem_id::ProblemGlobalId, PgLang, ProblemInfo, SampleTestcase, Url};
 
-use crate::client::SessionPersistentClient;
-use crate::config::{SubmissionConfig, TestConfig};
-use crate::interactive::ask_credential;
-use crate::storage::{
-    workspace, PlatformVault, ProblemVault, ProblemWorkspace, Repository, WorkspaceNameModifier,
+use self::error::*;
+use crate::{
+    client::SessionPersistentClient,
+    config::{SubmissionConfig, TestConfig},
+    interactive::{self, ask_credential},
+    storage::{
+        workspace, PlatformVault, ProblemVault, ProblemWorkspace, Repository, WorkspaceNameModifier,
+    },
+    style,
+    testing::{AsyncTestcase, FsTestcase, JudgeCode, TestOutcome, TestRunner},
 };
-use crate::style;
-use crate::testing::{AsyncTestcase, FsTestcase, JudgeCode, TestOutcome, TestRunner};
 
 pub async fn login(cli: &mut SessionPersistentClient) -> Result<()> {
     ensure!(
@@ -201,7 +198,7 @@ pub async fn create_contest_workspace(
                 },
             )
             .context("Failed to create contest workspace")?;
-        println!(
+        log::info!(
             "Successfully created workspace {}",
             loc.dir().to_string_lossy()
         );
@@ -243,9 +240,7 @@ pub async fn do_test(
         runner.compile().await?;
     }
 
-    let style = ProgressStyle::default_bar()
-        .template("{spinner} {msg}")
-        .unwrap();
+    let style = ProgressStyle::default_spinner();
 
     let mut results = Vec::with_capacity(testcases.len());
     let mut bars = Vec::with_capacity(testcases.len());
@@ -259,20 +254,8 @@ pub async fn do_test(
             .add(ProgressBar::new(100))
             .with_style(style.clone())
             .with_message(format!("Testcase {} ...", t.name()));
-        let bar = Arc::new(Mutex::new(bar));
-        bars.push(bar.clone());
-
-        // Tick spinner
-        tokio::spawn(async move {
-            loop {
-                tokio::time::sleep(Duration::from_millis(50)).await;
-                let bar = bar.lock().await;
-                if bar.is_finished() {
-                    break;
-                }
-                bar.tick();
-            }
-        });
+        let bar = interactive::tick_spinner(bar);
+        bars.push(bar);
     }
 
     for (t, bar) in testcases.iter().zip(&bars) {
@@ -301,122 +284,10 @@ pub async fn do_test(
     results
         .iter()
         .filter(|x| x.judge != JudgeCode::AC)
-        .for_each(print_test_result_detail);
+        .for_each(style::print_test_result_detail);
 
-    print_test_result_summary(&results);
+    style::print_test_result_summary(&results);
     Ok(results)
-}
-
-fn print_test_result_summary(results: &[TestOutcome]) {
-    let bar = "-".repeat(5);
-    print!("{} ", bar);
-
-    let count: HashMap<JudgeCode, usize> = results.iter().fold(HashMap::new(), |mut count, r| {
-        *count.entry(r.judge).or_default() += 1;
-        count
-    });
-
-    let num_total_test = results.len();
-    let num_passed = *count.get(&JudgeCode::AC).unwrap_or(&0);
-    let num_failed = num_total_test - num_passed;
-
-    if num_passed == num_total_test {
-        let msg = format!("All {} tests passed ✨", num_total_test);
-        print!("{}", msg.green());
-    } else {
-        let summary_msg = if num_passed > 0 {
-            format!("{}/{} tests failed 💣", num_failed, num_total_test)
-        } else {
-            format!("All {} tests failed 💀", num_total_test)
-        };
-
-        let detail_msg = count
-            .iter()
-            .filter(|(&judge, _)| judge != JudgeCode::AC)
-            .map(|(&judge, &cnt)| {
-                format!(
-                    "{}{}{}",
-                    style::judge_icon(judge),
-                    "x".dimmed(),
-                    cnt.to_string().bold().bright_white(),
-                )
-            })
-            .collect::<Vec<String>>()
-            .join(", ");
-
-        print!("{} ({})", summary_msg.bright_red(), detail_msg);
-    }
-
-    println!(" {}", bar);
-}
-
-pub fn print_test_result_detail(res: &TestOutcome) {
-    let stdout_lines: Vec<_> = res.output.stdout.lines().collect();
-    let truth_lines: Vec<_> = res.groundtruth.lines().collect();
-
-    let (cols, _) = terminal::size().unwrap_or((40, 40));
-
-    const BOLD_LINE: &str = "━";
-    const THIN_LINE: &str = "─";
-
-    let bold_bar = BOLD_LINE.repeat(cols as usize).blue().bold();
-
-    let title_color = Color::BrightYellow;
-    println!(
-        "\n{}: {} [{}ms]\n{}",
-        res.testcase_name.color(title_color).bold(),
-        style::judge_icon(res.judge),
-        res.execution_time.as_millis(),
-        bold_bar,
-    );
-
-    fn print_sub_title(s: &str, cols: usize) {
-        println!(
-            "{}{}",
-            s.cyan().bold(),
-            THIN_LINE.repeat(cols - s.len() - 1).bright_black(),
-        )
-    }
-
-    fn print_lines(lines: &[&str], entire_str: &str) {
-        if lines.is_empty() {
-            println!("{}", "<EMPTY>".magenta().dimmed());
-            return;
-        }
-        for (i, line) in lines.iter().enumerate() {
-            let trimmed = line.trim_end();
-            print!("{}", trimmed);
-
-            let num_trailling_whitespace = line.len() - trimmed.len();
-            if num_trailling_whitespace > 0 {
-                print!(
-                    "{}{}",
-                    " ".repeat(num_trailling_whitespace).on_red(),
-                    "(Trailling whitespace)".bright_red().bold()
-                );
-            }
-
-            let is_last_line = i + 1 == lines.len();
-            if is_last_line && !entire_str.ends_with("\n") {
-                print!("{}", " Missing new line ".on_yellow().black().bold());
-            }
-
-            println!("");
-        }
-    }
-
-    print_sub_title("[truth-answer]", cols as usize);
-    print_lines(&truth_lines, &res.groundtruth);
-
-    print_sub_title("[stdout]", cols as usize);
-    print_lines(&stdout_lines, &res.output.stdout);
-
-    if !res.output.stderr.is_empty() {
-        print_sub_title("[stderr]", cols as usize);
-        print!("{}", res.output.stderr);
-    }
-
-    println!("{}", bold_bar);
 }
 
 pub async fn submit(
