+++ core/src/interactive.rs
@@ -1,3 +1,7 @@
+use std::{sync::Arc, time::Duration};
+use tokio::sync::Mutex;
+
+use indicatif::ProgressBar;
 use kpr_webclient::{CredFieldKind, CredFieldMeta, CredMap};
 
 pub mod util {
@@ -36,4 +40,20 @@ pub fn ask_credential(fields: &[CredFieldMeta]) -> CredMap {
         map.insert(name, value);
     }
     map
+}
+
+pub fn tick_spinner(bar: ProgressBar) -> Arc<Mutex<ProgressBar>> {
+    let mutex_bar = Arc::new(Mutex::new(bar));
+    let bar = mutex_bar.clone();
+    tokio::spawn(async move {
+        loop {
+            tokio::time::sleep(Duration::from_millis(30)).await;
+            let bar = bar.lock().await;
+            if bar.is_finished() {
+                break;
+            }
+            bar.tick();
+        }
+    });
+    mutex_bar
 }
